revision.c: make default history consider bottom commits
Previously, the default history treated bottom commits the same as any
other UNINTERESTING commit, which could force it down side branches.

Consider the following history:

   *A--*B---D--*F         * marks !TREESAME parent paths
     \     /*
      `-C-'

When requesting "B..F", B is UNINTERESTING but TREESAME to D. C is
!UNINTERESTING.

So default following would go from D into the irrelevant side branch C
to A, rather than to B.  Note also that if there had been an extra
!UNINTERESTING commit B1 between B and D, it wouldn't have gone down C.

Change the default following to test relevant_commit() instead of
!UNINTERESTING, so it can proceed straight from D to B, thus finishing
the traversal of that path.

Signed-off-by: Kevin Bracey <kevin@bracey.fi>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: a5339c9036b6b1a7f7cf2bda9f2e75c05acc5bd6 [formerly a5339c9036b6b1a7f7cf2bda9f2e75c05acc5bd6 [formerly 141efdba57b1769fc60ff9a3925afbc6af398faf]]
Former-commit-id: 8aeef4ed200ed8a61ab5a60f0b8b7a0772c15473
Former-commit-id: 70c52aa0f7d5d48b24ececea95399526c0cde246