t5000: integrate export-subst tests into regular tests
Instead of creating extra archives for testing substitutions, set the
attribute export-subst and overwrite the marked file with the expected
(expanded) content right between committing and archiving.  Thus
placeholder expansion based on the committed content is performed with
each archive creation and the comparison with the contents of directory
a yields the correct result.  We can then remove the special tests for
export-subst.

Signed-off-by: René Scharfe <rene.scharfe@lsrfire.ath.cx>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 345944c03af5518bfbb90854a0115b4b8e902fd2 [formerly c6fce0f37ed34b414f680ff54ada7d65918f4410] [formerly c6fce0f37ed34b414f680ff54ada7d65918f4410 [formerly c420df7b9bbfcf352a3602c78e8476f18098bb43]]
Former-commit-id: 64eed2c1315e14a44bf44afa36b3f03e64c205db
Former-commit-id: 0135747b7f43ba78ae4cfae80a9bc588b7183a7f