t5000: test long filenames
Add a file with a long name to the test archive in order to check
entries with pax extended headers.  Also add a check for tar versions
that doen't understand this format.  Those versions should extract the
headers as a regular files.  Add code to check_tar() to interpret the
path header if present, so that our tests work even with those tar
versions.

It's important to use the fallback code only if needed to still be
able to detect git archive errorously creating pax headers as regular
file entries (with a suitable tar version, of course).

The archive used to check for pax header support in tar was generated
using GNU tar 1.26 and its option --format=pax.

Tested successfully on NetBSD 6.1, which has a tar version lacking pax
header support.

Signed-off-by: René Scharfe <rene.scharfe@lsrfire.ath.cx>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 0ddbd423c705f50ef56f627cabf1b8cd00b229f1 [formerly 0ddbd423c705f50ef56f627cabf1b8cd00b229f1 [formerly 9bf1ac41d215f9b1c026fe585a38ffa0555a5670]]
Former-commit-id: bdcef7516d2c7ac041a38c219eb37d924ea522b8
Former-commit-id: 7aac9b0a13c307ed3a95e40e15183fe77cce248e