remote-hg: always point HEAD to master
Mercurial always checks out the 'default' branch, so there's no point in
complicating our lives trying to do something fancier, which causes
different behavior depending on whether the repository is local or
remote.

So let's always use 'default' (which we translate to 'master'), unless
we are in hg-git mode, which expects us to use the 'master' bookmark
instead.

Also, update the tests that used to check for different checkout
behaviors to simply check that the refs are there, remove unnecessary
ones, and fix the ones that expect something different.

Signed-off-by: Felipe Contreras <felipe.contreras@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 584bc5fe27867fba5a88999c52f26640e472144b [formerly 584bc5fe27867fba5a88999c52f26640e472144b [formerly 91347ea3e17b853155f9cab31f0c2a28143284ed]]
Former-commit-id: 0096badeb1b4cf72522c53661a3fa8f5c3d1fba8
Former-commit-id: e67ed04bd0cff07ddf710b1bb9001bbf88ca0db7