remote-hg: implement custom push()
The one from mercurial does a ton of things we are not interested in,
and we need some special modifications which are impossible otherwise.

Most of the code is borrowed from Mercurial, and cleaned up, but should
be functionally the same for our purposes, except that multiple heads
are not detected.

Signed-off-by: Felipe Contreras <felipe.contreras@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: f033a09a5cd63d3a066ac6a2ada58fad830a1e98 [formerly 09febcbb770f4dfc52dcedd2565ac3745e45e35a] [formerly 09febcbb770f4dfc52dcedd2565ac3745e45e35a [formerly 4f37bdbdb6229931c464b99eb9f6e198de01d0a7]]
Former-commit-id: 3e6f9713e8b4a96bbdfb8aa7283b94648fa26147
Former-commit-id: 6db53148908452fbdca25d93b3b1deac031ececf