send-email: remove warning about unset chainreplyto
Three years and a half is probably more than enough time to give users
the opportunity to configure Git to do what they want. If they haven't
changed the configuration by now, this warning message is not going to
do anything for them anyway.

This effectively reverts commit 528fb08 (prepare send-email for smoother
change of --chain-reply-to default).

Signed-off-by: Felipe Contreras <felipe.contreras@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: ab9113e1a10423be0de9152d91ef7ae553cdc017 [formerly dc5916c2010236b8a727525d61a017585298ee95] [formerly dc5916c2010236b8a727525d61a017585298ee95 [formerly b99d22f29a58feeccb1eb129b68a73f974715ee1]]
Former-commit-id: fcdc7ecc1d0caf11a993a8bcf0f2ceac20a85c23
Former-commit-id: 82f2e4c13dd64dcda2e707297c0d4b699094fd0a