object_array_remove_duplicates(): rewrite to reduce copying
The old version copied one entry to its destination position, then
deleted any matching entries from the tail of the array.  This
required the tail of the array to be copied multiple times.  It didn't
affect the complexity of the algorithm because the whole tail has to
be searched through anyway.  But all the copying was unnecessary.

Instead, check for the existence of an entry with the same name in the
*head* of the list before copying an entry to its final position.
This way each entry has to be copied at most one time.

Extract a helper function contains_name() to do a bit of the work.

Signed-off-by: Michael Haggerty <mhagger@alum.mit.edu>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: d85a28853ca7d897cabad82d6c345aee958b386f [formerly d85a28853ca7d897cabad82d6c345aee958b386f [formerly 1506510c170d23b8f769757dc81904f334c40281]]
Former-commit-id: 1ebd56e9969d3d7bacf5fa98f960388f7f766c5a
Former-commit-id: e388a24253e06b0992f1d441d92489a2224767bf