refs: document the lifetime of the args passed to each_ref_fn
The lifetime of the memory pointed to by the refname and sha1
arguments to each_ref_fn was never documented, but some callers used
to assume that it was essentially permanent.  In fact the API does
*not* guarantee that these objects live beyond a single callback
invocation.

In the current code, the lifetimes are bound together with the
lifetimes of the ref_caches.  Since these are usually long, the
callers usually got away with their sloppiness.  But even today, if a
ref_cache is invalidated the memory can be freed.  And planned changes
to reference caching, needed to eliminate race conditions, will
probably need to shorten the lifetimes of these objects.

The commits leading up to this have (hopefully) fixed all of the
callers of the for_each_ref()-like functions.  This commit does the
last step: documents what each_ref_fn callbacks can assume about
object lifetimes.

Signed-off-by: Michael Haggerty <mhagger@alum.mit.edu>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 1cab3f02b34fef5bfaede879d75be95aacc5d84d [formerly 8a3c6914379174a08adce04665d57f448027787d] [formerly 8a3c6914379174a08adce04665d57f448027787d [formerly 4f78c24c63bf0b035afc02372727a3b5897d9835]]
Former-commit-id: 02e13dbdfec2cc3282656ce76c701430afec345d
Former-commit-id: 911406900dd529a3122be77ef66f80c2d620ccb5