prune-packed: avoid implying "1" is DRY_RUN in prune_packed_objects()
Commit b60daf0 (Make git-prune-packed a bit more chatty. - 2007-01-12)
changes the meaning of prune_packed_objects()'s argument, from "dry
run or not dry run" to a bitmap.

It however forgot to update prune_packed_objects() caller in
builtin/prune.c to use new DRY_RUN macro. It's fine (for a long time!)
but there is a risk that someday someone may change the value of
DRY_RUN to something else and builtin/prune.c suddenly breaks. Avoid
that possibility.

While at there, change "opts == VERBOSE" to "opts & VERBOSE" as there
is no obvious reason why we only be chatty when DRY_RUN is not set.

Signed-off-by: Nguyễn Thái Ngọc Duy <pclouds@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 7bb0ec42fd4ba0c2c1994539b22eed32b48f72cf [formerly 7bb0ec42fd4ba0c2c1994539b22eed32b48f72cf [formerly af0b4a3b5973ebdf8fcdedcccbc2e8ddb93c8534]]
Former-commit-id: ef8fa525d7d8b4e2798e0d1a2d7ca16df964eec4
Former-commit-id: e6080659e388daa5f8ebd1bdda8fb653ffa52bbc