fix segfault with git log -c --follow
In diff_tree_combined we make a copy of diffopts. In
try_to_follow_renames, called via diff_tree_sha1, we free and
re-initialize diffopts->pathspec->items. Since we did not make a deep
copy of diffopts in diff_tree_combined, the original diffopts does not
get the update. By the time we return from diff_tree_combined,
rev->diffopt->pathspec->items points to an invalid memory address. We
get a segfault next time we try to access that pathspec.

Instead, along with the copy of diffopts, make a copy pathspec->items as
well.

We would also have to make a copy of pathspec->raw to keep it consistent
with pathspec->items, but nobody seems to rely on that.

Signed-off-by: Clemens Buchacher <drizzd@aon.at>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 6671ae4cc64c48306a5e5098b7441d6b8924fb40 [formerly 6671ae4cc64c48306a5e5098b7441d6b8924fb40 [formerly 46ec510ac088ac2669e617a4c8c35e6218dabecc]]
Former-commit-id: e9d7a6d7250bd4281f48ad783862be0f3b1bf50b
Former-commit-id: ac555b44c92783e72860b85cecd85c4b74e65e2a