git-remote-mediawiki: better error message when HTTP(S) access fails
My use-case is an invalid SSL certificate. Pulling from the wiki with a
recent version of libwww-perl fails, and git-remote-mediawiki gave no
clue about the reason. Give the mediawiki API detailed error message, and
since it is not so informative, hint the user about an invalid SSL
certificate on https:// urls.

Signed-off-by: Matthieu Moy <Matthieu.Moy@imag.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 4af64c35b5333485af2702104ec9ace3d316c48a [formerly 4af64c35b5333485af2702104ec9ace3d316c48a [formerly 0aa0321212d6b817783616af4ba1353a46b82179]]
Former-commit-id: 9fefa0461c8af2f5532b4a9b8d62aa432d00ddaa
Former-commit-id: 84760f88498c4e7942840ee9f58c78a28810bbce