get_sha1: warn about full or short object names that look like refs
When we get 40 hex digits, we immediately assume it's an SHA-1. This
is the right thing to do because we have no way else to specify an
object. If there is a ref with the same object name, it will be
ignored. Warn the user about this case because the ref with full
object name is likely a mistake, for example

    git checkout -b $empty_var $(git rev-parse something)

advice.object_name_warning is not documented because frankly people
should not be aware about it until they encounter this situation.

While at there, warn about ambiguation with abbreviated SHA-1 too.

Signed-off-by: Nguyễn Thái Ngọc Duy <pclouds@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 5d27ca126b6a9385491d4e080ef8c6a2e52acca2 [formerly 5d27ca126b6a9385491d4e080ef8c6a2e52acca2 [formerly 798c35fcd8a71a094ca68ac05d81e08c5ac8166d]]
Former-commit-id: d1192e33161b80772e00c8873859027ac6a6c4ac
Former-commit-id: 564e3ffb3b6aa67f1609ac56bab9d8d8dd5ad803