difftool --dir-diff: allow changing any clean working tree file
The temporary directory prepared by "difftool --dir-diff" to
show the result of a change can be modified by the user via
the tree diff program, and we try hard not to lose changes
to them after tree diff program returns to us.

However, the set of files to be copied back is computed
differently between --symlinks and --no-symlinks modes.  The
former checks all paths that start out as identical to the
working tree file, while the latter checks paths that
already had a local modification in the working tree,
allowing changes made in the tree diff program to paths that
did not have any local change to be lost.

Signed-off-by: Kenichi Saita <nitoyon@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 0d7d3a00f7ed73a6805cbe12b99a24b5d8f5bdde [formerly 0d7d3a00f7ed73a6805cbe12b99a24b5d8f5bdde [formerly 32eaf1de7f79c4ba09f3de3261c84e52e0a67af5]]
Former-commit-id: 20437b51c2c138f24d599e1773fce0bc30161297
Former-commit-id: a09c1a98e4a8a8a2a57cbe28bed2a38b9397200f