push: make push.default = current use resolved HEAD
With this change, the output of the push (with push.default set to
current) changes subtly from:

  $ git push
  ...
   * [new branch]      HEAD -> push-current-head

to:

  $ git push
  ...
   * [new branch]      push-current-head -> push-current-head

This patch was written with a different motivation. There is a problem
unique to push.default = current:

  # on branch push-current-head
  $ git push
  # on another terminal
  $ git checkout master
  # return to the first terminal
  # the push tried to push master!

This happens because the 'git checkout' on the second terminal races
with the 'git push' on the first terminal.  Although this patch does not
solve the core problem (there is still no guarantee that 'git push' on
the first terminal will resolve HEAD before 'git checkout' changes HEAD
on the second), it works in practice.

Signed-off-by: Ramkumar Ramachandra <artagnon@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: efa47aaeb6311bf7f0f6da49e830b0e0183a3f5e [formerly 99ad4b5d30c33a4d66cc9fb614ecbc6325805d1d] [formerly 99ad4b5d30c33a4d66cc9fb614ecbc6325805d1d [formerly 0f075b220862e729eb3873a1c1496b923d17c2d4]]
Former-commit-id: 6f68ad6db97bb23bdb389f52639d2a188194ac3f
Former-commit-id: cc3bb2b4bf1d5c3355e3fe2bd61c5a0f01a66b82