path: Fix a sparse warning
On MinGW, sparse issues an "'get_st_mode_bits' not declared. Should
it be static?" warning. The MinGW and MSVC builds do not see the
declaration of this function, within git-compat-util.h, due to its
placement within an preprocessor conditional.

In order to suppress the warning, we simply move the declaration to
the top level of the header.

Signed-off-by: Ramsay Jones <ramsay@ramsay1.demon.co.uk>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: b3c1b19709dfbca25cd11a724c888b739ce61a26 [formerly b3c1b19709dfbca25cd11a724c888b739ce61a26 [formerly 2f0aaaf9dabd846061a759b53ff4759ec886df62]]
Former-commit-id: dd276a63b598e99d4dc4d7c42f6b8f826819251d
Former-commit-id: 130783c3aae06cc2e3095b79e0d86c7aac26295f