wildmatch: properly fold case everywhere
Case folding is not done correctly when matching against the [:upper:]
character class and uppercased character ranges (e.g. A-Z).
Specifically, an uppercase letter fails to match against any of them
when case folding is requested because plain characters in the pattern
and the whole string are preemptively lowercased to handle the base case
fast.

That optimization is kept and ISLOWER() is used in the [:upper:] case
when case folding is requested, while matching against a character range
is retried with toupper() if the character was lowercase, as the bounds
of the range itself cannot be modified (in a case-insensitive context,
[A-_] is not equivalent to [a-_]).

Signed-off-by: Anthony Ramine <n.oxyde@gmail.com>
Reviewed-by: Duy Nguyen <pclouds@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 8c02a8e486017474da918bcccaede3722bc031e9 [formerly 17598a4ad77ce6caaa1793be1844caf03c888b11] [formerly 17598a4ad77ce6caaa1793be1844caf03c888b11 [formerly b79c0c3755f3694a3c7ecd3cad18bda011283db7]]
Former-commit-id: 6cc46a9e21d322975a6f303d81b253a9d665a266
Former-commit-id: d6f8b527c76dfb1008ffc717738762c871008aee