sha1_file: silence sha1_loose_object_info
sha1_object_info() returns -1 (OBJ_BAD) if it cannot find the object
for some reason, which suggests that it wants the _caller_ to report
this error.  However, part of its work happens in
sha1_loose_object_info, which _does_ report errors itself.  This is
doubly strange because:

* packed_object_info(), which is the other half of the duo, does _not_
  report this.

* In the event that an object is packed and pruned while
  sha1_object_info_extended() goes looking for it, we would
  erroneously show the error -- even though the code of the latter
  function purports to handle this case gracefully.

* A caller might invoke sha1_object_info() to find the type of an
  object even if that object is not known to exist.

Silence this error.  The others remain untouched as a corrupt object
is a much more grave error than it merely being absent.

Signed-off-by: Thomas Rast <trast@inf.ethz.ch>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: a83b1045f2fb66608fc1baac80df051ec902f83d [formerly a83b1045f2fb66608fc1baac80df051ec902f83d [formerly dbea72a8c05d12c7dfd6a35376e62b21d5a655b5]]
Former-commit-id: 056a1e9e4359f48ded331aa18c0331571936afae
Former-commit-id: fb3a4763a4eece2ed9bc11e1b45b45cd78f18513