Test 'commit --only' after 'checkout --orphan'
There are some index handling subtleties in 'commit --only' that are
best tested when we have an existing index, but an unborn or empty
HEAD.  These circumstances are easily produced by 'checkout --orphan',
but we did not previously have a test for it.

The main expected failure mode would be: erroneously loading the
existing index contents when building the temporary index that is used
for --only.  Cf.

  http://article.gmane.org/gmane.comp.version-control.git/225969

and subsequent discussion.

Signed-off-by: Thomas Rast <trast@inf.ethz.ch>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: e88d101fbc52d72fc5a02d39e184eeb3ce0bda12 [formerly e88d101fbc52d72fc5a02d39e184eeb3ce0bda12 [formerly f0c73200e86b6166596f45ce32b7f04291074182]]
Former-commit-id: 16e4354548c83038597171379e756b9b3460fd26
Former-commit-id: 3d227723e960d06923524a7380555b11226bd30f