completion: difftool takes both revs and files
'git difftool' is clearly a frontend to 'git diff' and is used in
exactly the same way, but it uses a misleadingly named completion
function __git_complete_file.  It happens to work only because it
calls __git_complete_revlist_file that completes both revs and
paths.

Change it to use __git_complete_revlist_file, just like 'git diff'.

Signed-off-by: Ramkumar Ramachandra <artagnon@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 64c7de0469bc833c0d83830f76c4e4fbb6fa69a5 [formerly 64c7de0469bc833c0d83830f76c4e4fbb6fa69a5 [formerly d8517cc6670d4ba18e9b77a83b2481aec0edbc32]]
Former-commit-id: 8f781eb571562bbce2a02f39a8d3c3675e0e5269
Former-commit-id: 8bc0d87a43347a146293b7aac6d158dccffc30a8