completion: show can take both revlist and paths
The 'git show' completion uses __git_complete_file (aliased to
__git_complete_revlist_file), because accepts <tree-ish>:<path> as
well as <commit-ish>.  But the command also accepts range of commits
in A..B notation, so using __git_complete_revlist_file is more
appropriate.

There still remain two users of __git_complete_file, completions for
"archive" and "ls-tree".  As these commands do not take range
notation, and "git show" no longer uses __git_complete_file, the
implementation of it can be updated not to complete ranges, but that
is a separate topic.

Signed-off-by: Ramkumar Ramachandra <artagnon@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: ca87082742ecea1b994eab5afbdfce8bab6b89e0 [formerly 9d7960e91263f3c6ece3e761b0974054279788ca] [formerly 9d7960e91263f3c6ece3e761b0974054279788ca [formerly 5269f7f8c6b293f1568c4f03c14596835e22df31]]
Former-commit-id: 475e28950c4338f9b1fd7146f2226ba1022dce24
Former-commit-id: c17cddc095274c3a4002752ac190c7e9f5172fb0