unpack-trees: create working copy of merge entry in merged_entry
Duplicate the merge entry right away and work with that instead of
modifying the entry we got and duplicating it only at the end of
the function.  Then mark that pointer const to document that we
don't modify the referenced cache_entry.

This change is safe because all existing merge functions call
merged_entry just before returning (or not at all), i.e. they don't
care about changes to the referenced cache_entry after the call.
unpack_nondirectories and unpack_index_entry, which call the merge
functions through call_unpack_fn, aren't interested in such changes
neither.

The change complicates merged_entry a bit because we have to free the
copy if we error out, but allows callers to pass a const pointer.

Signed-off-by: René Scharfe <rene.scharfe@lsrfire.ath.cx>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 7b982755862c232d6cf863c322246918f014c02c [formerly cf56e7c5de1bddb8d11dc5d1c41b106d4114396c] [formerly cf56e7c5de1bddb8d11dc5d1c41b106d4114396c [formerly f2fa35420511cc49e85413a2932a1a2bac88cc1b]]
Former-commit-id: acd5c9506d6cbccbd4a081b711d8027da03373d9
Former-commit-id: 9fd2557706141816e879624fc7d13630317c636e