diff-lib, read-tree, unpack-trees: mark cache_entry array paramters const
Change the type merge_fn_t to accept the array of cache_entry pointers
as const pointers to const pointers.  This documents the fact that the
merge functions don't modify the cache_entry contents or replace any of
the pointers in the array.

Only a single cast is necessary in unpack_nondirectories because adding
two const modifiers at once is not allowed in C.  The cast is safe in
that it doesn't mask any modfication; call_unpack_fn only needs the
array for reading.

Signed-off-by: René Scharfe <rene.scharfe@lsrfire.ath.cx>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: bb9e5bb8a958c4c995776408cadde3aac20d0774 [formerly bb9e5bb8a958c4c995776408cadde3aac20d0774 [formerly 5828e8352c07753a1f751322800524bf9dff8679]]
Former-commit-id: 71cc27ec994dfd4207e085106e294dff7d1a9969
Former-commit-id: 75e5e75dcd57dc00fd20f3fad5094cbcbcc6347f