t/README: test_must_fail is for testing Git
When a test wants to make sure there is no <string> in an output
file, we should just say "! grep string output".

"test_must_fail" is there only to test Git command and catch unusual
deaths we know about (e.g. segv) as an error, not as an expected
failure.  "test_must_fail grep string output" is unnecessary, as
we are not making sure the system binaries do not dump core or
anything like that.

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 6d95218ad8b25770c2181a7657f30dfa0d7aa056 [formerly 6d95218ad8b25770c2181a7657f30dfa0d7aa056 [formerly f445500e4dd79ad0c4f6cf1c4949f01f155142f7]]
Former-commit-id: daaa0458c9e2d56e40cb05dad37ef91dc7346bff
Former-commit-id: 0c8a380b258f07988b1b54b0712af02912b03339