gitweb: fix problem causing erroneous project list
The bug is manifest when running gitweb in a persistent process (e.g.
FastCGI, PSGI), and it's easy to reproduce.  If a gitweb request
includes the searchtext parameter (i.e. s), subsequent requests using
the project_list action--which is the default action--and without
a searchtext parameter will be filtered by the searchtext value of the
first request.  This is because the value of the $search_regexp global
(the value of which is based on the searchtext parameter) is currently
being persisted between requests.

Instead, clear $search_regexp before dispatching each request.

Signed-off-by: Charles McGarvey <chazmcgarvey@brokenzipper.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 9fce439d931eb06fcc6a030177f9be0fc522cf73 [formerly 9fce439d931eb06fcc6a030177f9be0fc522cf73 [formerly ca7a5dcfd39bdeb26d74074e78e6966df5e63a6b]]
Former-commit-id: 2aeed0561a407c3b28b1a7edc99bea1a2447ba52
Former-commit-id: bf36736097cd8702c1e73f6bfecffe4a48bdd4be