t0005: skip signal death exit code test on Windows
The test case depends on that test-sigchain can commit suicide by a
call to raise(SIGTERM) in a way that run-command.c::wait_or_whine()
can detect as death through a signal. There are no POSIX signals on
Windows, and a sufficiently close emulation is not available in the
Microsoft C runtime (and probably not even possible).

The particular deficiency is that when a signal is raise()d whose
SIG_DFL action will cause process death (SIGTERM in this case), the
implementation of raise() in msvcrt just calls exit(3).

We could check for exit code 3 in addition to 143, but that would
miss the point of the test entirely. Hence, just skip it on Windows.

Signed-off-by: Johannes Sixt <j6t@kdbg.org>
Acked-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: ec68d70303c387237ad22c8df0d10fadadcf3eed [formerly ec68d70303c387237ad22c8df0d10fadadcf3eed [formerly 04422c74c88e2d5d78210bad65e95b223a18fd8e]]
Former-commit-id: 6ccee726787af53284f72ddf34b2919d0db27055
Former-commit-id: 191a34de3dff9aaddf32600951fad3e1764595d4