prio-queue: priority queue of pointers to structs
Traditionally we used a singly linked list of commits to hold a set
of in-flight commits while traversing history.  The most typical use
of the list is to add commits that are newly discovered to it, keep
the list sorted by commit timestamp, pick up the newest one from the
list, and keep digging.  The cost of keeping the singly linked list
sorted is nontrivial, and this typical use pattern better matches a
priority queue.

Introduce a prio-queue structure, that can be used either as a LIFO
stack, or a priority queue.  This will be used in the next patch to
hold in-flight commits during sort-in-topological-order.

Tests and the idea to make it usable for any "void *" pointers to
"things" are by Jeff King.  Bugs are mine.

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: e1937ec0d07f843a21c7d4ceaa7c1f9bdc8fd9d3 [formerly e1937ec0d07f843a21c7d4ceaa7c1f9bdc8fd9d3 [formerly b4b594a3154078430b04fad4f6ffbed9c7274be5]]
Former-commit-id: 700c95c95e851af8f2e4ac32be5643b3a8ae8e97
Former-commit-id: b85a34a08bad1b480d43bccc2381fd2cd2a74b50