fetch-options.txt: prevent a wildcard refspec from getting misformatted
When explaining the "--tags" option as an equivalent to giving an
explicit "refs/tags/*:refs/tags/*" refspec, the two asterisks were
misinterpreted by AsciiDoc as a request to typeset the string
segment between them in bold.

We could fix it in two ways.  We can replace them with {asterisk}s
while keeping the string as body text, or we can mark it as a
literal string with backquotes around it.

Let's do the latter, as it is teaching the user an "exactly as
typed" alternative.

Noticed-by: Ramkumar Ramachandra <artagnon@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: b05703c56ef5ff048e48546a155b655d4bf79a56 [formerly b05703c56ef5ff048e48546a155b655d4bf79a56 [formerly 9eb4754d7682aa10c6dcbcc503a50a716d9cfc8a]]
Former-commit-id: 3d4648695e6fe73f0a213f7134b39cd44c03f703
Former-commit-id: 9a1787b644bf9916406153ae1a9f200735b30e4a