t3030: use test_ln_s_add to remove SYMLINKS prerequisite
The test cases include many corner-cases of merge-recursive's behavior,
some of them involve type changes and symbolic links. All cases, including
those that are protected by SYMLINKS check only whether the result of
merge-recursive is correctly stored in the database and the index; the
file system is not investigated. Use test_ln_s_add to enter a symbolic
link in the index in the test setup and run the tests without the
SYMLINKS prerequisite.

Notice that one test that has the SYMLINKS protection removed is an
expect_failure. There is a possibility that the test fails differently
depending on whether SYMLINKS is present or not; but this is not the case
presently.

Signed-off-by: Johannes Sixt <j6t@kdbg.org>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: ff6ab0dfe4ed76bb82223b7030b910b26f36a415 [formerly ff6ab0dfe4ed76bb82223b7030b910b26f36a415 [formerly bba56042e7011a5555cb8b44a84e0d853fa43af5]]
Former-commit-id: 026abd343adab0b91a0f04431add4f9f9373245d
Former-commit-id: 7f8f3e5cc4c92e3fe661afbf4e440ad917cc99dc