t6035: use test_ln_s_add to remove SYMLINKS prerequisite
All tests in t6035 are protected by SYMLINKS. But that is not necessary,
because a lot of the functionality can be tested provided symbolic link
entries enter the index and object data base. Use test_ln_s_add for this
purpose.

Some test cases do test the presence of symbolic links on the file system.
Move these tests into separate test cases that remain protected by
SYMLINKS.

There is one instance of expect_failure. There is a possibility that this
test case fails differently depending on whether SYMLINKS is present or
not; but this is not the case.

Signed-off-by: Johannes Sixt <j6t@kdbg.org>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: a5920f137d2b147f8f752d14162c92faeba8f966 [formerly a5920f137d2b147f8f752d14162c92faeba8f966 [formerly bfd7804ee3d5285aa8598497995efe3f697ea7fd]]
Former-commit-id: 6ea471d9191f71ed7f547141e71a6ae896ca2d4b
Former-commit-id: ef6c1d311edc47393284fe3145a138ca47fa9f64