t4011: remove SYMLINKS prerequisite
The part of the test that is about symbolic links in the index does not
require that the corresponding file system entry is actually a symbolic
link. Use test_ln_s_add to insert a symbolic link in the index. When
the file system does not support symbolic links, we actually have a
regular file in the worktree, which  we can update as if it were a
symbolic link. diff-index picks up the symbolic link property from the
index.

Signed-off-by: Johannes Sixt <j6t@kdbg.org>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 844a920c5558d29f49d9d52455a35a9aa5b0ed19 [formerly 5db76770fa920dbe5fadee36381381855e31be33] [formerly 5db76770fa920dbe5fadee36381381855e31be33 [formerly 70836a6f8b71a945f074fc796c4241c2183b83a2]]
Former-commit-id: 083587ca377cf14d7c633cbfb9157d26ee6019ca
Former-commit-id: a513a23dd4fc2ed4167bf1ad6773aa9df09dd845