build: generate and clean test scripts
Commit 416fda6 (build: do not install git-remote-testpy) made it so
git-remote-testpy is not only not installed, but also not generated
by default.  From a fresh checkout, "make --test=5800 test" would
have failed.

This was not found primarily because "make clean" failed to remove
git-remote-testpy, which is another bug in the same commit.

Fix the former by having 'all' target depend on $(NO_INSTALL) and
the latter by removing $(NO_INSTALL) in the 'clean' target.

Signed-off-by: Felipe Contreras <felipe.contreras@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 78068583d4faff8dba8c7aef1d3d355571c9e422 [formerly 78068583d4faff8dba8c7aef1d3d355571c9e422 [formerly 6c473a56d26bf596d4314b7c2e4c2bfafbfe757e]]
Former-commit-id: 854488662c289e8b78590503bc4284980feb67e1
Former-commit-id: 75d4b7158c8d3cb71c3e9ee086d6daeb29f46288