t/lib-httpd/apache.conf: configure an MPM module for apache 2.4
Versions of Apache before 2.4 always had a "MultiProcessing
Module" (MPM) statically built in, which manages the worker
threads/processes. We do not care which one, as it is
largely a performance issue, and we put only a light load on
the server during our testing.

As of Apache 2.4, the MPM module is loadable just like any
other module, but exactly one such module must be loaded. On
a system where the MPMs are compiled dynamically (e.g.,
Debian unstable), this means that our test Apache server
will not start unless we provide the appropriate
configuration.

Unfortunately, we do not actually know which MPM modules are
available or appropriate for the system on which the tests
are running. This patch picks the "prefork" module, as it
is likely to be available on all Unix-like systems.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 1fb1b240f2915d7b6a745e425ce01c3f5fe6dafb [formerly edb52261f0d5bc07e8dca4ddca2aa475307c2255] [formerly edb52261f0d5bc07e8dca4ddca2aa475307c2255 [formerly 296f0b3ea9b904982780522215a9c71d4a7bd02c]]
Former-commit-id: fcf76486de882c757daf51bc85d43e5fcece2dba
Former-commit-id: 39df3d59a931bd97b38f1f4978e63c632bbe7d43