test: test_must_be_empty helper
There are quite a lot places where an output file is expected to be
empty, and we fail the test when it is not.  The output from running
the test script with -i -v can be helped if we showed the unexpected
contents at that point.

We could of course do

    >expected.empty && test_cmp expected.empty actual

but this is commmon enough to be done with a dedicated helper.

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: dd4c9de5aafdd2598f129610300491fae8240346 [formerly dd4c9de5aafdd2598f129610300491fae8240346 [formerly ca8d148daf3014577222c2562ca2c8170a866aa4]]
Former-commit-id: f5a49cc6a0c9b304452d1458378dd2f3716262c4
Former-commit-id: 2bcba1a999d3309377bfefa18d2e3f2d475546f7