Move copy_note_for_rewrite + friends from builtin/notes.c to notes-utils.c
This is a pure code movement of the machinery for copying notes to
rewritten objects. This code was located in builtin/notes.c for
historical reasons. In order to make it available to builtin/commit.c
it was declared in builtin.h. This was more of an accident of history
than a concious design, and we now want to make this machinery more
widely available.

Hence, this patch moves the code into the new notes-utils.[hc] files
which are included into libgit.a. Except for adjusting #includes
accordingly, this patch merely moves the relevant functions verbatim
into the new files.

Cc: Thomas Rast <trast@inf.ethz.ch>
Signed-off-by: Johan Herland <johan@herland.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 69e97e9e89ea8749d8b4be03d2eecfce4ed285ca [formerly 69e97e9e89ea8749d8b4be03d2eecfce4ed285ca [formerly 49c2470400bbb695b953abeb4e6737be747317af]]
Former-commit-id: 730daf39a5e68537c1f0e9d422161a9748fd8392
Former-commit-id: 0c0a4a86079e834f395747ec343de74d56a4bb3c