add--interactive: respect diff.algorithm
When staging hunks interactively it is sometimes useful to use an
alternative diff algorithm which splits the changes into hunks in a more
logical manner.  This is not possible because the plumbing commands
called by add--interactive ignore the "diff.algorithm" configuration
option (as they should).

Since add--interactive is a porcelain command it should respect this
configuration variable.  To do this, make it read diff.algorithm and
pass its value to the underlying diff-index and diff-files invocations.

At this point, do not add options to "git add", "git reset" or "git
checkout" (all of which can call git-add--interactive).  If a user
wants to override the value on the command line they can use:

	git -c diff.algorithm=$ALGO ...

Signed-off-by: John Keeping <john@keeping.me.uk>
Reviewed-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: e64624c25470c7a839f7852fe84da9a5b916f9a3 [formerly e64624c25470c7a839f7852fe84da9a5b916f9a3 [formerly 2cc0f53b53e7d5456ec45713a201142754d69249]]
Former-commit-id: cc8a0e10a7bcf0f500d59a1936bf548be37201e5
Former-commit-id: 5be55a8dcd86f17c9df57ac7a184247f6287db3c