contrib: drop blameview/ directory
Blameview was a quick-and-dirty demonstration of how blame's
incremental output could be used in an interface. These days
one can find much better (and less ugly!) demonstrations in
"git gui blame" and "tig blame".

The only advantage blameview has is that its code is perhaps
simpler to read. However, that is balanced by the fact that
it probably has bugs, as nobody uses it nor has touched the
code in 6 years. An implementor is probably better off just
reading the "incremental output" section of "man git-blame".

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: e7a3ba15421c5b9fb06f6cd34f78dcd9fb1b7d6c [formerly e7a3ba15421c5b9fb06f6cd34f78dcd9fb1b7d6c [formerly 1af6a877c5aeeb97e8fbacb06992e976e5cdbf83]]
Former-commit-id: e272bbf573572d4914d0e7ad7af7f9bf4a171f74
Former-commit-id: 1e474c5841345f328e415055b96011e74ae6828c