use logical OR (||) instead of binary OR (|) in logical context
The compiler can short-circuit the evaluation of conditions strung
together with logical OR operators instead of computing the resulting
bitmask with binary ORs.  More importantly, this patch makes the
intent of the changed code clearer, because the logical context (as
opposed to binary context) becomes immediately obvious.

While we're at it, simplify the check for patch->is_rename in
builtin/apply.c a bit; it can only be 0 or 1, so we don't need a
comparison operator.

Signed-off-by: René Scharfe <rene.scharfe@lsrfire.ath.cx>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 577ca592872a3cc6559a9cbe44ed67fce41cfd96 [formerly 577ca592872a3cc6559a9cbe44ed67fce41cfd96 [formerly 0b437a18bdd188794cec8855e2a413921d08b6e8]]
Former-commit-id: 02a8b8b6c3b589b247fa9317b3c782db2158ec4b
Former-commit-id: e86f9fc0e061adc572ee80e61f52ad2ab949dcb0