handle multibyte characters in name
Many "git submodule" operations do not work on a submodule at a path whose
name is not in ASCII.

This is because "git ls-files" is used to find which paths are bound to
submodules to the current working tree, and the output is C-quoted by default
for non ASCII pathnames.

Tell "git ls-files" to not C-quote its output, which is easier than unwrapping
C-quote ourselves.

Signed-off-by: Fredrik Gustafsson <iveqy@iveqy.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 9fff244a7a68bd76cce451f22ef87303b21b1f07 [formerly 9fff244a7a68bd76cce451f22ef87303b21b1f07 [formerly 74671241fd53df2899d4b4400593afc12e8e1bf7]]
Former-commit-id: a0a3418afd8f24037fafe7c686b187ea93d9dcae
Former-commit-id: 342671257a6ec80cdfc5f11e39f15ed8737d4afe