git-remote-mediawiki: make a regexp clearer
Perl's split function takes a regex pattern argument. You can also
feed it an expression, which is then compiled into a regex at runtime.
It therefore works to pass your pattern via single quotes, but it is
much less obvious to a reader that the argument is meant to be a
regex, not a static string. Using the traditional slash-delimiters
makes this easier to read.

Signed-off-by: Célestin Matte <celestin.matte@ensimag.fr>
Signed-off-by: Matthieu Moy <matthieu.moy@grenoble-inp.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 32eeaa6bbe602acc23dc5582272b3729608d8114 [formerly cf3aead6f50ae87b92684164179ddcdad70b968a] [formerly cf3aead6f50ae87b92684164179ddcdad70b968a [formerly 05d4c7b1c46b7a6cbc8f4c0403a6e208aea1c63c]]
Former-commit-id: a17a762103b58f42154a9d96846a63735a8ca741
Former-commit-id: cc04cc105d4e681eb063360c5cbffd598d46b1d4