git-remote-mediawiki: change the behaviour of a split
A "split ' '" is turned into a "split / /", which changes its behaviour: the
old method matched a run of whitespaces (/\s*/), while the new one will match a
single space, which is what we want here. Indeed, in other contexts,
changing split(' ') to split(/ /) could potentially be a regression, however,
here, when parsing the output of "rev-list --parents", whose output SHA-1's are
each separated by a single space, splitting on a single space is perfectly
correct.

Signed-off-by: Célestin Matte <celestin.matte@ensimag.fr>
Signed-off-by: Matthieu Moy <matthieu.moy@grenoble-inp.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 6edb61775f663c512bbfc1d3315817e8906c59fd [formerly 6edb61775f663c512bbfc1d3315817e8906c59fd [formerly 1149957368ef29b6a05ad5cf93f6c62352136bcd]]
Former-commit-id: 7e0479356d97dd2a19dbef3e6cd7ee843afa7a7f
Former-commit-id: cb11a063d52c5fcbbe2d34c6db94eef943238abc