stash: introduce 'git stash store'
save_stash() contains the logic for doing two potentially independent
operations; the first is preparing the stash merge commit, and the
second is updating the stash ref/ reflog accordingly.  While the first
operation is abstracted out into a create_stash() for callers to access
via 'git stash create', the second one is not.  Fix this by factoring
out the logic for storing the stash into a store_stash() that callers
can access via 'git stash store'.

Like create, store is not intended for end user interactive use, but for
callers in other scripts.  We can simplify the logic in the
rebase.autostash feature using this new subcommand.

Signed-off-by: Ramkumar Ramachandra <artagnon@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 25f67f9912b71b6b7bcee75e6eb32f0d4e8e5124 [formerly 25f67f9912b71b6b7bcee75e6eb32f0d4e8e5124 [formerly bd514cada4bb1a396a58d408c3ec08526a8be742]]
Former-commit-id: 7a235f08bd00a1d4709109dc19c4fbefe38c5279
Former-commit-id: 4aa8b0c26410061ec755f9b3c888fd459e2495b5