unpack-trees: don't shift conflicts left and right
If o->merge is set, the struct traverse_info member conflicts is shifted
left in unpack_callback, then passed through traverse_trees_recursive
to unpack_nondirectories, where it is shifted right before use.  Stop
the shifting and just pass the conflict bit mask as is.  Rename the
member to df_conflicts to prove that it isn't used anywhere else.

Signed-off-by: René Scharfe <rene.scharfe@lsrfire.ath.cx>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 8b8c6dcc00d68a9922193426f206d5668888c1d9 [formerly 8b8c6dcc00d68a9922193426f206d5668888c1d9 [formerly 603d249853981d81d1f8cf561b18d77cb1b55923]]
Former-commit-id: f190f844c894adc2fbb761a78671461d6f8fba8a
Former-commit-id: c7180d17b32b6ff3f2a11e75f861f3850fb84233