t/t2021-checkout-last: "checkout -" should work after a rebase finishes
  $ git checkout -

does not work as expected after a rebase.  This is because the
reflog records "checkout" made by "rebase" as its implementation
detail the same way as end-user initiated "checkout", and makes it
count as the branch that was previously checked out.

Add four failing tests documenting this bug: two for a normal rebase,
and another two for an interactive rebase.

Signed-off-by: Ramkumar Ramachandra <artagnon@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 71862c7c6974bbd6d9b2c113ec0441752db36ed6 [formerly 71862c7c6974bbd6d9b2c113ec0441752db36ed6 [formerly 89f2fea49a246e9ac4fa94cb43a7e992a8a4a144]]
Former-commit-id: cb19194b6e7fdb9e58936e749dd4aa76796a4f1e
Former-commit-id: ffc6429e1b699b3a8da6a87c6a6271157e80e56a