rebase -i: use a better reflog message
Now that the "checkout" invoked internally from "rebase -i" knows to
honor GIT_REFLOG_ACTION, we can start to use it to write a better
reflog message when "rebase anotherbranch", "rebase --onto branch",
etc. internally checks out the new fork point.  We will write:

  rebase -i: checkout master

instead of the old

  rebase -i

As all the calls git-rebase--interactive make to underlying git
commands that leave reflog messages are preceded by the internal
comment_for_reflog helper function, which uses the original value of
the GIT_REFLOG_ACTION variable it saw when it first started, the new
assignments to GIT_REFLOG_ACTION actively contaminate the value of
the variable, knowing that it will be reset to a sane value before
it is used again.  This does not generally hold true but it should
suffice for now.

Signed-off-by: Ramkumar Ramachandra <artagnon@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 3824af0f124511bd80f4b0d9705acc5be7c52866 [formerly 15d4d2005d8f69b03934d3ec38fcd97cc3c94e99] [formerly 15d4d2005d8f69b03934d3ec38fcd97cc3c94e99 [formerly 26cd160cb15635afd1c2937a3c15cfb256758323]]
Former-commit-id: 6fc5436ab23819993406ddf19742697556c1b08d
Former-commit-id: 1806ae4fbb23478c986f92e8053d20d8f14e50ce