rev-parse: add --prefix option
This makes 'git rev-parse' behave as if it were invoked from the
specified subdirectory of a repository, with the difference that any
file paths which it prints are prefixed with the full path from the top
of the working tree.

This is useful for shell scripts where we may want to cd to the top of
the working tree but need to handle relative paths given by the user on
the command line.

Signed-off-by: John Keeping <john@keeping.me.uk>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 0842bdd988e23fbd7557268b98ebdf4578841376 [formerly 0842bdd988e23fbd7557268b98ebdf4578841376 [formerly 12b9d32790b40bf3ea49134095619700191abf1f]]
Former-commit-id: c705b32973d12f18a8848c381d32539f08f2b1d0
Former-commit-id: 95abe00e52398635c2572d84473bd5e738eb7d77