submodule: drop the top-level requirement
Use the new rev-parse --prefix option to process all paths given to the
submodule command, dropping the requirement that it be run from the
top-level of the repository.

Since the interpretation of a relative submodule URL depends on whether
or not "remote.origin.url" is configured, explicitly block relative URLs
in "git submodule add" when not at the top level of the working tree.

Signed-off-by: John Keeping <john@keeping.me.uk>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: aa79cf521619139c285969a0ab080da7360504f1 [formerly aa79cf521619139c285969a0ab080da7360504f1 [formerly 091a6eb0feed820a43663ca63dc2bc0bb247bbae]]
Former-commit-id: 183a29a83c184beca5cbb67576483d1b5cef7cda
Former-commit-id: 04ce9498bb52f2ac357b748cec66b534bb36b444