completion, bash prompt: move __gitdir() tests to completion test suite
Currently __gitdir() is duplicated in the git completion and prompt
scripts, while its tests are in the prompt test suite.  This patch
series is about to change __git_ps1() in a way that it won't need
__gitdir() anymore and __gitdir() will be removed from the prompt
script.

So move all __gitdir() tests from the prompt test suite over to the
completion test suite.  Update the setup tests so that they perform
only those steps that are necessary for each test suite.

Signed-off-by: SZEDER Gábor <szeder@ira.uka.de>

Former-commit-id: 6f838e69bcd535d4c007a81b14f03d4b720890c4 [formerly a2173162c059e0a0d4a93f1744c2f23e533fcd57] [formerly a2173162c059e0a0d4a93f1744c2f23e533fcd57 [formerly c9a102e81ffedde3fed4b88199ea13a3a5ee5f5d]]
Former-commit-id: e9937441ca1182a1cd03676648c8b28dc08826f5
Former-commit-id: 341a2b54985ff6e5221bfe6bfd4fb964325fc57d