bash prompt: combine 'git rev-parse' executions in the main code path
There are a couple of '$(git rev-parse --<opt>)' command substitutions
in __git_ps1() and three of them are executed in the main code path:

 - the first to get the path to the .git directory ('--git-dir'),
 - the second to check whether we're inside the .git directory
   ('--is-inside-git-dir'),
 - and the last, depending on the results of the second, either
   * to check whether it's a bare repo ('--is-bare-repository'), or
   * to check whether inside a work tree ('--is-inside-work-tree').

Naturally, this imposes the overhead of fork()ing three subshells and
fork()+exec()ing three git commands.

Combine these four 'git rev-parse' queries into a single one and use
bash parameter expansions to parse the combined output, i.e. to
separate the path to the .git directory from the true/false of
'--is-inside-git-dir', etc.  This way we can eliminate two of the
three subshells and git commands.

Signed-off-by: SZEDER Gábor <szeder@ira.uka.de>

Former-commit-id: 7c686453509105c200fbacb7ad1e026752ba7a53 [formerly 7c686453509105c200fbacb7ad1e026752ba7a53 [formerly efaa0c153297f551a42fd1e21f28f51f4924f316]]
Former-commit-id: 0cb478ee03933ce6f47e956473c3cddf58387a2e
Former-commit-id: a9eb55c24fcd46aa012f6ea2249415b3508a31ac