builtin/checkout.c: don't leak memory in check_tracking_name
remote_find_tracking() populates the query struct with an allocated
string in the dst member.  So, we do not need to xstrdup() the string,
since we can transfer ownership from the query struct (which will go
out of scope at the end of this function) to our callback struct, but
we must free the string if it will not be used so we will not leak
memory.

Let's do so.

Signed-off-by: Brandon Casey <drafnel@gmail.com>
Reviewed-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: ac0910a7f566f428367c3ec4bc51c62f4cb5c122 [formerly ac0910a7f566f428367c3ec4bc51c62f4cb5c122 [formerly 2c4842023205b4b2cf7eac973cd7dc2fca263327]]
Former-commit-id: cf351f7a12073d61594f29c46e5bd5b46488c0fa
Former-commit-id: c26b388e46dbbe3379416b779c299922fd0973d6