t/t9802: explicitly name the upstream branch to use as a base
Prior to commit fa83a33b, the 'git checkout' DWIMery would create a
new local branch if the specified branch name did not exist and it
matched exactly one ref in the "remotes" namespace.  It searched
the "remotes" namespace for matching refs using a simple comparison
of the trailing portion of the remote ref names.  This approach
could sometimes produce false positives or negatives.

Since fa83a33b, the DWIMery more strictly excludes the remote name
from the ref comparison by iterating through the remotes that are
configured in the .gitconfig file.  This has the side-effect that
any refs that exist in the "remotes" namespace, but do not match
the destination side of any remote refspec, will not be used by
the DWIMery.

This change in behavior breaks the tests in t9802 which relied on
the old behavior of searching all refs in the remotes namespace,
since the git-p4 script does not configure any remotes in the
.gitconfig.  Let's work around this in these tests by explicitly
naming the upstream branch to base the new local branch on when
calling 'git checkout'.

Signed-off-by: Brandon Casey <drafnel@gmail.com>
Acked-by: Pete Wyckoff <pw@padd.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: dc7a88822a0e39bdf6bac81c71d1cf4fe8cad116 [formerly dc7a88822a0e39bdf6bac81c71d1cf4fe8cad116 [formerly 9d58c4a3e3ce3a40262b7ff3cc39ce3b9a93d527]]
Former-commit-id: 62f4cdf949de39fcea7c61e595fc84bca1b525c7
Former-commit-id: 891b6ac0bf089fe3e5f7dcaff877bc412a1dd6c1