rebase topology tests: fix commit names on case-insensitive file systems
The recently introduced tests used uppercase letters to denote
cherry-picks of commits having the corresponding lowercase letter names.
The helper functions also set up tags with the names of the commits.

But this constellation fails on case-insensitive file systems because
there cannot be distinct tags with names that differ only in case.

Use a less subtle convention for the names of cherry-picked commits.

Signed-off-by: Johannes Sixt <j6t@kdbg.org>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 84942154daace8260c346efb43193526dd95bac5 [formerly 21b51a685083611c21370b435bfa68d67d0b75f3] [formerly 21b51a685083611c21370b435bfa68d67d0b75f3 [formerly 984f78d278dd456b12874da1bdf3763e72db8dae]]
Former-commit-id: af94b0b13a224b8d272e3448377a2d8a8955839e
Former-commit-id: 683e3236ccc91cea80242e658d09eb5d0a95d814