Documentation/git-merge.txt: weaken warning about uncommited changes
Commit 35d2fffd introduced 'git merge --abort' as a synonym to 'git reset
--merge', and added some failing tests in t7611-merge-abort.sh (search
'###' in this file) showing that 'git merge --abort' could not always
recover the pre-merge state.

Still, in many cases, 'git merge --abort' just works, and it is usually
considered that the ability to start a merge with uncommited changes is
an important property of Git.

Weaken the warning by discouraging only merge with /non-trivial/
uncommited changes.

Signed-off-by: Matthieu Moy <Matthieu.Moy@imag.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 8ed11fe9a16ff51301307ee952dbda04fa50a856 [formerly 8ed11fe9a16ff51301307ee952dbda04fa50a856 [formerly 76b80cdf17807ba4a773562df2218980ffa27e4f]]
Former-commit-id: 108f67a6a2481d54161aefc642653ef4a6f38584
Former-commit-id: 5da0d30ca846b1ebb0bc85e7dcb8e319157e66ec