name-rev: allow to specify a subpath for --refs option
When an user wants to filter specific ref using the --refs option,
the pattern needs to match the full ref, e.g. --refs=refs/tags/v1.*.

It'd be convenient to specify a subpath of ref pattern.  For
example, --refs=origin/* can find refs/remotes/origin/master by
searching the pattern against its substrings in turn:

  refs/remotes/origin/master
  remotes/origin/master
  origin/master

If it finds a match in a subpath, unambigous part of the ref path will
be removed in the output.

Signed-off-by: Namhyung Kim <namhyung.kim@lge.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 1756d976f12a979d95884e8d274689e2871fc9cf [formerly 1756d976f12a979d95884e8d274689e2871fc9cf [formerly 98c5c4ad01551b7764c9c8d8f67abfa9c5e595fe]]
Former-commit-id: 89a95deb4643afbec03fa7ac85bfcac1cfff2750
Former-commit-id: c5ff2aab3ce5f82f8367b3dbc4256a69b8111736