Documentation: Move "git diff <blob> <blob>"
The section describing "git diff <blob> <blob>" had been placed in a
position that disrupted the statement "This is synonymous to the
previous form".

Reorder to place this form after all the <commit>-using forms, and the
note applying to them. Also mention this form in the initial description
paragraph.

Signed-off-by: Kevin Bracey <kevin@bracey.fi>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: cfa008c4f5dae7da7e41590d8a0deca9dd104100 [formerly f07c93de252a1d962016a503fe07676584c039cb] [formerly f07c93de252a1d962016a503fe07676584c039cb [formerly d7747bd5bcba5eaeda690d7d5897462580f3c0e4]]
Former-commit-id: c8fc0f05a076a0f28ff77f90253fb1a4a0f7c890
Former-commit-id: 8e3c2c770e72e1370e4c65dc00a40cf6c1e12491