http.c: don't rewrite the user:passwd string multiple times
Curl older than 7.17 (RHEL 4.X provides 7.12 and RHEL 5.X provides
7.15) requires that we manage any strings that we pass to it as
pointers.  So, we really shouldn't be modifying this strbuf after we
have passed it to curl.

Our interaction with curl is currently safe (before or after this
patch) since the pointer that is passed to curl is never invalidated;
it is repeatedly rewritten with the same sequence of characters but
the strbuf functions never need to allocate a larger string, so the
same memory buffer is reused.

This "guarantee" of safety is somewhat subtle and could be overlooked
by someone who may want to add a more complex handling of the username
and password.  So, let's stop modifying this strbuf after we have
passed it to curl, but also leave a note to describe the assumptions
that have been made about username/password lifetime and to draw
attention to the code.

Signed-off-by: Brandon Casey <drafnel@gmail.com>
Acked-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 7d71a2bf24951c1d1b6628187bce2aca31d8ae90 [formerly 490962ddc86c8800314edd272853a6850b7afa2b] [formerly 490962ddc86c8800314edd272853a6850b7afa2b [formerly a94cf2cb7e77c27ce5c3d648e37c1aa75cd5e56e]]
Former-commit-id: 5a6fbdb85f0785958baeb5c36ac86ad2701115bf
Former-commit-id: d9d15f31ec54e18abc322123392d5a0b323131ac