resolve_ref_unsafe(): handle the case of an SHA-1 within loop
There is only one "break" statement within the loop, which jumps to
the code after the loop that handles the case of a file that holds a
SHA-1.  So move that code from below the loop into the if statement
where the break was previously located.  This makes the logic flow
more local.

Signed-off-by: Michael Haggerty <mhagger@alum.mit.edu>
Reviewed-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 8f154f5d0ce7c32b08202e9fd4f94cd02388c4ce [formerly 8f154f5d0ce7c32b08202e9fd4f94cd02388c4ce [formerly 2884c06ae78e569a513701c2d43f4ed79bd252ce]]
Former-commit-id: bb4d98634b8a8e1aa43337a4034e38d077f38de2
Former-commit-id: 1395011ad68b0c9e3ebb1cab70c021e024cfdf6f