config doc: rewrite push.default section
4d35924e (Merge branch 'rr/triangle', 2013-04-07) introduced support
for triangular workflows, but the push.default values still assume
central workflows.

Rewrite the descriptions of `nothing`, `current`, `upstream` and
`matching` for greater clarity, and explicitly explain how they
should behave in triangular workflows.

Leave `simple` as it is for the moment, as we plan to change its
meaning to accommodate triangular workflows in a later patch.

Signed-off-by: Ramkumar Ramachandra <artagnon@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: f4e5c6e9afca1914d72b7755261225f35dc5db62 [formerly f4e5c6e9afca1914d72b7755261225f35dc5db62 [formerly 87a70e4ce8bb3bdbb3048a5eb837f6b5b2eff8f9]]
Former-commit-id: cdf4db9087fc6313c02b41c1034713f93a67f3ec
Former-commit-id: e8e5d5d2dbf2c3ea31f56bc82728db38d0578cb9