t/t5528-push-default: generalize test_push_*
The setup creates two bare repositories: repo1 and repo2, but
test_push_commit() hard-codes checking in repo1 for the actual output.
Generalize it and its caller, test_push_success(), to optionally accept
a third argument to specify the name of the repository to check for
actual output.  We will use this in the next patch.

Signed-off-by: Ramkumar Ramachandra <artagnon@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: c22b0e41cd759680a4cd1a77e0125efebc8b8f8a [formerly e5722a573777526b9efe3898e5eb88f78bedfa10] [formerly e5722a573777526b9efe3898e5eb88f78bedfa10 [formerly 396243fa479cd27bef1ced2ea5a9e4cf20f85a4b]]
Former-commit-id: 5703bd708f5e2c82d26785659ddf4830341f0da3
Former-commit-id: c331570663bae0fed2189b2d255ce971caa5e18a