refs: manage lifetime of packed refs cache via reference counting
In struct packed_ref_cache, keep a count of the number of users of the
data structure.  Only free the packed ref cache when the reference
count goes to zero rather than when the packed ref cache is cleared.
This mechanism will be used to prevent the cache data structure from
being freed while it is being iterated over.

So far, only the reference in struct ref_cache::packed is counted;
other users will be adjusted in separate commits.

Signed-off-by: Michael Haggerty <mhagger@alum.mit.edu>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 88bf16456c999296820fd6a3fc62c51d0cb0bc0c [formerly 88bf16456c999296820fd6a3fc62c51d0cb0bc0c [formerly 5f5e2a8868e2b5792bf467799339e3288282e91e]]
Former-commit-id: 3184ca92c95474dd8cdd3aa724835c874a808d76
Former-commit-id: 360888bc58446e9dde5f3d7feeaa44da9221f89c