add a stat_validity struct
It can sometimes be useful to know whether a path in the
filesystem has been updated without going to the work of
opening and re-reading its content. We trust the stat()
information on disk already to handle index updates, and we
can use the same trick here.

This patch introduces a "stat_validity" struct which
encapsulates the concept of checking the stat-freshness of a
file. It is implemented on top of "struct stat_data" to
reuse the logic about which stat entries to trust for a
particular platform, but hides the complexity behind two
simple functions: check and update.

Signed-off-by: Michael Haggerty <mhagger@alum.mit.edu>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 408fd1d23474e59f44ba120c9ce09efa84b2b025 [formerly 408fd1d23474e59f44ba120c9ce09efa84b2b025 [formerly 38612532240ecbe1b12e54ca859fed8410ae6de1]]
Former-commit-id: 396bf8bdbd28d3170d84cadb307c43bd921dd912
Former-commit-id: 999edfbb8561e7838a3a4823f55b0a216ce54213