refs: do not invalidate the packed-refs cache unnecessarily
Now that we keep track of the packed-refs file metadata, we can detect
when the packed-refs file has been modified since we last read it, and
we do so automatically every time that get_packed_ref_cache() is
called.  So there is no need to invalidate the cache automatically
when lock_packed_refs() is called; usually the old copy will still be
valid.

Signed-off-by: Michael Haggerty <mhagger@alum.mit.edu>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 07d582e56c1d9d257f6295b841043ca03be2a97e [formerly 07d582e56c1d9d257f6295b841043ca03be2a97e [formerly 5d478f5ca1792e7576ac13aaf75f09099ce99afb]]
Former-commit-id: 3069f0b7f029685fcce3ddb1c470522a4a1a381e
Former-commit-id: a40fc7bf27c4fccb5cc084f7e10fb8b61708796f