transport-helper: be quiet on read errors from helpers
Prior to commit 81d340d4, we did not print any error message
if a remote transport helper died unexpectedly. If a helper
did not print any error message (e.g., because it crashed),
the user could be left confused. That commit tried to
rectify the situation by printing a note that the helper
exited unexpectedly.

However, this makes a much more common case worse: when a
helper does die with a useful message, we print the extra
"Reading from 'git-remote-foo failed" message. This can also
end up confusing users, as they may not even know what
remote helpers are (e.g., the fact that http support comes
through git-remote-https is purely an implementation detail
that most users do not know or care about).

Since we do not have a good way of knowing whether the
helper printed a useful error, and since the common failure
mode is for it to do so, let's default to remaining quiet.
Debuggers can dig further by setting GIT_TRANSPORT_HELPER_DEBUG.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: a278860b456ce26cf9cf4f49ab53ec13bf7cb52b [formerly d12a1d7933f950d7f3f9dc528bfc3f3c59a154be] [formerly d12a1d7933f950d7f3f9dc528bfc3f3c59a154be [formerly 266f1fdfa99f5d29ca7ce455966e7960c00a82e4]]
Former-commit-id: 947015b4e707b67dc4dbb0a77d3776d160cc98e6
Former-commit-id: b671ef74d839218939151e02a3e9ade1ae891d8d