apply: carefully strdup a possibly-NULL name
2901bbe (apply: free patch->{def,old,new}_name fields, 2012-03-21)
cleaned up the memory management of filenames in the patches, but
forgot that find_name_traditional() can return NULL as a way of saying
"I couldn't find a name".

That NULL unfortunately gets passed into xstrdup() next, resulting in
a segfault.  Use null_strdup() so as to safely propagate the null,
which will let us emit the correct error message.

Reported-by: DevHC on #git
Signed-off-by: Thomas Rast <trast@inf.ethz.ch>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 6ac7bba6045007024824edf07a36c97920135e52 [formerly 6ac7bba6045007024824edf07a36c97920135e52 [formerly 212eb96a96e256933a76e48d85fa4f500bed0dee]]
Former-commit-id: 4ce3cec2a45a88fe0e1fc2f8610644befd0f5318
Former-commit-id: 3eb3987e52a4d3ffd7122f661b20948587119270