topology tests: teach a helper to take abbreviated timestamps
The on_committer_date helper in t/lib-t6000 is used in t6002 and
t6003 with timestamps on a single day within a single minute
(i.e. 1971-08-16 00:00) and the tests repeat this over and over.

The actual value of the timestamp, however, does not matter very
much; only their relative ordering does.

Introduce another helper to expand only the suffix of the timestamp
to a full timestamp to make the lines shorter, and use it in this
helper.  Also, because all the commits in the test are made with
specific GIT_COMMITTER_DATE, stop unsetting it at the end of the
helper.

We'll be specifying the author timestamp to these test commits in a
later patch, which will be helped with this change.

Also remove a test that was commented-out from t6003; it used to
test a commit with the same parent listed twice, which was allowed
by mistake but was fixed long time ago.

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 9486d4ae61bedc4559914bc4a747c98ec0e58c83 [formerly 9486d4ae61bedc4559914bc4a747c98ec0e58c83 [formerly 841dc6935ea0f2ecfc778501edc66dc2c828d333]]
Former-commit-id: 2301d5a211baa2e770dc068dc17cb28dc1b12844
Former-commit-id: e0475fb75901fc0421cb4f5ba868323173fa91b7