lib-httpd/apache.conf: check version only after mod_version loads
Commit 0442743 introduced an <IfVersion> directive near the
top of the apache config file. However, at that point we
have not yet checked for and loaded the mod_version module.
This means that the directive will behave oddly if
mod_version is dynamically loaded, failing to match when it
should.

We can fix this by moving the whole block below the
LoadModule directive for mod_version.

Reported-by: Brian Gernhardt <brian@gernhardtsoftware.com>
Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: f466cc6216bfc73f69fab580e9aca54bc6d3f3e5 [formerly f466cc6216bfc73f69fab580e9aca54bc6d3f3e5 [formerly 5bdc47eb55aabef3720039181fc360499081a29f]]
Former-commit-id: c9a7013a3849cb24daafd55061e34dbbf26a8ba9
Former-commit-id: 46c74810e82a0f3cc29e836e6012300d05cbca5c