remote: remove dead code in read_branches_file()
The first line of the function checks that the remote-name contains a
slash ('/'), and sets the "slash" variable accordingly.  The only caller
of read_branches_file() is remote_get_1(); the calling codepath is
guarded by valid_remote_nick(), which checks that the remote does not
contain a slash.  Therefore, the "slash" variable can never be set:
remove the dead code that assumes otherwise.

Signed-off-by: Ramkumar Ramachandra <artagnon@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 8c8c202bc294abaac5dd61a05872c403564dd39e [formerly 8c8c202bc294abaac5dd61a05872c403564dd39e [formerly 55cfde251b4bee36f64f2067cccc4ba346aaf1b0]]
Former-commit-id: fa79c0a542ce5eea5aa83abeed8d674891716320
Former-commit-id: 87831283e2205923bc39a348fc1a7b6705b5d7a5