ls-remote doc: fix example invocation on git.git
Under the EXAMPLES section, there is one invocation on the git.git
repository that attempts to list the refs master, pu, and rc.  The ref
rc does not exist in today's repository, so remove it.  Among other
things, this example demonstrates that the "<refs>..." argument is
simply a filter; requesting a non-existent ref is not an error.

Signed-off-by: Ramkumar Ramachandra <artagnon@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: b242f1d5c4e03328b8cb9eb2c76d50e107067f2e [formerly f9f888a592ace65b9094682bf47cac0bd5fb5a7c] [formerly f9f888a592ace65b9094682bf47cac0bd5fb5a7c [formerly 6077d3629964d0ba03798a82fd69b745574142ed]]
Former-commit-id: acb7d5a980652397b8d51354b5d0096d1ae674bb
Former-commit-id: 18e5464f5ced2c73fafd0a41eb39ad5012c2c386