completion: handle unstuck form of base git options
git-completion.bash's parsing of the command name relies on everything
preceding it starting with '-' unless it is the "-c" option.  This
allows users to use the stuck form of "--work-tree=<path>" and
"--namespace=<path>" but not the unstuck forms "--work-tree <path>" and
"--namespace <path>".  Fix this.

Similarly, the completion only handles the stuck form "--git-dir=<path>"
and not "--git-dir <path>", so fix this as well.

Signed-off-by: John Keeping <john@keeping.me.uk>
Acked-by: SZEDER Gábor <szeder@ira.uka.de>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 5ee39e07e8bb3665b771f43a32833b436cc36d89 [formerly 5ee39e07e8bb3665b771f43a32833b436cc36d89 [formerly 776009d1b2bcfbd001767758fea0c2ff0e74383d]]
Former-commit-id: 6fc9f5c98d2700eea965206a6bb69c303173cb10
Former-commit-id: cc081726da6e7206bc45dff6083f961cb6c163df