doc/clone: Pick more compelling paths for the --reference example
There may be times when using one of your local repositories as a
reference for a new clone make sense, but the implied version-bump in
the old example isn't one of them.  I think a more intuitive example
is multi-user system with a central reference clone, and the new paths
hint at this use case.

Signed-off-by: W. Trevor King <wking@tremily.us>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: ebe567fccb8b0d289f47f5360852bcfe11037ed0 [formerly 9c2b7c7d7a137445dec9a28fb1223c5a7f43f041] [formerly 9c2b7c7d7a137445dec9a28fb1223c5a7f43f041 [formerly f22a6543d17d615068652d0f94397f558158d91a]]
Former-commit-id: 3d1b352b1111c7bfd74d2a1a43594b1d3ca9f197
Former-commit-id: 40c4929d2eb831fe7a07d83773abf9f3b8f988fd