fast-import: allow moving the root tree
Because fast-import.c::tree_content_remove does not check for the empty
path, it is not possible to move the root tree to a subdirectory.
Instead the error "Path  not in branch" is produced (note the double
space where the empty path has been inserted).

Fix this by explicitly checking for the empty path and handling it.

Signed-off-by: John Keeping <john@keeping.me.uk>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: a213b8c503ea90ea79c8e4d2a092389946126b54 [formerly 4a7ad711b333dbbff043c9f2ac62e065da794a8c] [formerly 4a7ad711b333dbbff043c9f2ac62e065da794a8c [formerly 62bfa11cc9a8cb0aacbba0aed503e85cdc6ba0a4]]
Former-commit-id: 35fbd623363a8b71d7c8cb0111e2d80dd972d045
Former-commit-id: 90feddd39ec6943823f936e81cc145fdfb6425b1