test-lib: self-test that --verbose works
t0000 contains some light self-tests of test-lib.sh, but --verbose was
not covered.  Add a test.

The only catch is that the presence of a test harness influences the
output (specifically, the presence of some empty lines).  So we need
to unset TEST_HARNESS or set it to a known value.  Leaving it unset
leads to spurious test failures in the final summary, which come from
the subtest.  So we always set it.

Signed-off-by: Thomas Rast <trast@inf.ethz.ch>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 5e24a6ad49b392f19a2963e9140d6f7ceeeecd18 [formerly 5e24a6ad49b392f19a2963e9140d6f7ceeeecd18 [formerly 517cd55fd51ebf4c6c20597055ff191fefe5dda0]]
Former-commit-id: 748cb1869313aece86c05a48879a0fa380fce4ce
Former-commit-id: 5befbd469cad0d0eda7f919c44a28a0e37d0889c