test-lib: verbose mode for only tests matching a pattern
With the new --verbose-only=<pattern> option, one can enable --verbose
at a per-test granularity.  The pattern is matched against the test
number, e.g.

  ./t0000-basic.sh --verbose-only='2[0-2]'

to see only the full output of test 20-22, while showing the rest in the
one-liner format.

As suggested by Jeff King, this takes care to wrap the entire
test_expect_* block, but nothing else, in the verbose toggling.  We
can use the test_start/end functions from the previous commit for the
purpose.

This is arguably not *too* useful on its own, but makes the next patch
easier to follow.

Helped-by: Jeff King <peff@peff.net>
Signed-off-by: Thomas Rast <trast@inf.ethz.ch>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 0e2ee88d706eef2156fedb8b5409cc656f454db1 [formerly 197c0a1b4c05b2dd29e76e39290f2654766227e9] [formerly 197c0a1b4c05b2dd29e76e39290f2654766227e9 [formerly ff09af3fb8f5bede0de523723bbd00d6ef2ab71e]]
Former-commit-id: 04865a6f54c6a0d0064dc8ed0fd8cadc483e13f4
Former-commit-id: 537d49605b044e72f5abe29c1fe28ebd1611b674