test-lib: allow prefixing a custom string before "ok N" etc.
This is not really meant for external use, and thus not documented. It
allows the next commit to neatly distinguish between sub-tests and the
main run.

The format is intentionally not valid TAP.  The use in the next commit
would not result in anything valid either way, and it seems better to
make it obvious.

Signed-off-by: Thomas Rast <trast@inf.ethz.ch>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 4fcbea5f962932a33de746703fb67c5af70d4a89 [formerly c0772f022d80cf98c21d90c609cb2cab872d0a7e] [formerly c0772f022d80cf98c21d90c609cb2cab872d0a7e [formerly e939e15d241e942662b9f88f6127ab470ab0a0b9]]
Former-commit-id: 463655629b5478825fe7ff3fbd8139219610ac07
Former-commit-id: 67f392a87411f6fc65989557825d011ba582ffb4