add -i: add extra options at the right place in "diff" command line
Appending "--diff-algorithm=histogram" at the end of canned command
line for various modes of "diff" is correct for most of them but not
for "stash" that has a non-option already wired in, like so:

	'stash' => {
		DIFF => 'diff-index -p HEAD',

Appending an extra option after non-option may happen to work due to
overly lax command line parser, but that is not something we should
rely on.  Instead, splice in the extra argument immediately after the
command name (i.e. 'diff-index', 'diff-files', etc.).

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 8cd0cc8ce15de30aa6496db2151c98c243893cf2 [formerly 9cd2e2a21d0112b72975e5e20a4eaf65d7cef28f] [formerly 9cd2e2a21d0112b72975e5e20a4eaf65d7cef28f [formerly e5c29097826be86e2f18139491cdf9c830c7fa3e]]
Former-commit-id: 7d93e380a74e739a61e048c041b7f9b8b7341b78
Former-commit-id: 59467dbba98dd203c3d948cc9f0c611b1fcb5ff1