commit: make it work with status.short
With "status.short" set, it is now impossible to commit with
status.short set, because it acts like "git commit --short", and it
is impossible to differentiate between a status_format set by the
command-line option parser versus that set by the config parser.

To alleviate this problem, clear status_format as soon as the config
parser has finished its work.

Signed-off-by: Ramkumar Ramachandra <artagnon@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: be70030a9d94bbc05a179847dca7229a7aed8ddc [formerly 27b37192e4b955d88497a8a2164fca51faaefea3] [formerly 27b37192e4b955d88497a8a2164fca51faaefea3 [formerly f0915cbaf476d63f72c284057680809ed24fbe0d]]
Former-commit-id: fca3f631d4a13ae4f670181bb14e42f4f306acd9
Former-commit-id: cc882a81bd71ddb647ccc3f7bdab22b50ae25ca0