write_name{_quoted_relative,}(): remove redundant parameters
After substitute path_relative() in quote.c with relative_path()
from path.c, parameters (such as len and prefix_len) are redundant
in function write_name() and write_name_quoted_relative().  The
callers have already been audited that the strings they pass are
properly NUL terminated and the length they give are the length of
the string (or -1 that asks the length to be counted by the callee).

Remove these now-redundant parameters.

Signed-off-by: Jiang Xin <worldhello.net@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 91ceefaf02d8c4c9d85a45f1e03ce818423e67ea [formerly 91ceefaf02d8c4c9d85a45f1e03ce818423e67ea [formerly e9a820cefde2170840fbcdf7c4b74369988869dc]]
Former-commit-id: fa46f5b676d9d2bc22c49584cd699df26056bbab
Former-commit-id: a99bfe51815a1668332fe7679be0d58e85cf957b