git-clean: add filter by pattern interactive action
Add a new action for interactive git-clean: filter by pattern. When the
user chooses this action, user can input space-separated patterns (the
same syntax as gitignore), and each clean candidate that matches with
one of the patterns will be excluded from cleaning. When the user feels
it's OK, presses ENTER and backs to the confirmation dialog.

Signed-off-by: Jiang Xin <worldhello.net@gmail.com>
Suggested-by: Junio C Hamano <gitster@pobox.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: ec392dcfa44f75fb1608e4160038683a96839d5f [formerly ec392dcfa44f75fb1608e4160038683a96839d5f [formerly d1239264f2786d7ea15543b980c6bf6afbc4701a]]
Former-commit-id: 543bc15a94587bd60887f50dc733dde31e31a51e
Former-commit-id: bb87c7518fc722ecc4ce14ecafeb7bea6c7055d8