git-prompt.sh: add missing information in comments
Mention that the command below is needed for prompt
in ZSH with PS1:
  setopt PROMPT_SUBST

Rephrase some parts that mention only the "current branch name"
being displayed in the prompt. Replace it by stating that
the "repository status" is displayed.

Make it clear that colored prompt is only available
in PROMPT_COMMAND/precmd mode.

With-suggestions-by: SZEDER Gábor <szeder@ira.uka.de>
Signed-off-by: Eduardo R. D'Avila <erdavila@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 0c9a626f725ade700edc9c4ea6cea82f0392838e [formerly 0c9a626f725ade700edc9c4ea6cea82f0392838e [formerly cf4cac4cfc13326920bdbf3d51236e2065f5ba2f]]
Former-commit-id: 2860e21d93424da7a42ac9c5714c8e1cebfe20cc
Former-commit-id: 4c82c873b05b3432eb4c0bddda3b3ca1b2828aca