documentation: add git:// transport security notice
The fact that the git:// transport does no authentication is easily
overlooked.  For example, DNS poisoning may result in fetching from
somewhere that was not intended.

Add a brief security notice to the "GIT URLS" section
of the documentation stating that the git transport should be used
with caution on unsecured networks.

Signed-off-by: Fraser Tweedale <frase@frase.id.au>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 4ce3c22bcdb35508f5e71f50f7dd82d58024dd5c [formerly f64efe24a009a8099239bfc3a0e3b556f638b091] [formerly f64efe24a009a8099239bfc3a0e3b556f638b091 [formerly 20618016df1255a2adebe900af7d66f88ef7a33b]]
Former-commit-id: 0d2391fbea092b64ab5c6dc5ee018592f6817893
Former-commit-id: 4ccd1eae30a5eac3b73c9ca4789da13ba7df57a4