rebase -i: handle fixup! fixup! in --autosquash
In rebase -i --autosquash, ignore all "fixup! " or "squash! " after the
first.  This supports the case when a git commit --fixup/--squash referred
to an earlier fixup/squash instead of the original commit (whether
intentionally, as when the user expressly meant to note that the commit
fixes an earlier fixup; or inadvertently, as when the user meant to refer to
the original commit with :/msg; or out of laziness, as when the user could
remember how to refer to the fixup but not the original).

In the todo list, the full commit message is preserved, in case it provides
useful cues to the user.  A test helper set_cat_todo_editor is introduced to
check this.

Helped-by: Thomas Rast <trast@inf.ethz.ch>
Helped-by: Junio C Hamano <gitster@pobox.com>
Signed-off-by: Andrew Pimlott <andrew@pimlott.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: f414d3cf58d07b46dc03b72c9055a666a6544daa [formerly f414d3cf58d07b46dc03b72c9055a666a6544daa [formerly 22c5b136363c7aa427667876e787db832548a038]]
Former-commit-id: 00ae0f20b5ca19ea92d11e41f8bcda20e19cd4de
Former-commit-id: 468101f1f6d3b658fc14348ffbb6357ad1eb9ef4