git stash: avoid data loss when "git stash save" kills a directory
"stash save" is about saving the local change to the working tree,
but also about restoring the state of the last commit to the working
tree.  When a local change is to turn a non-directory to a directory,
in order to restore the non-directory, everything in the directory
needs to be removed.

Which is fine when running "git stash save --include-untracked",
but without that option, untracked, newly created files in the
directory will have to be discarded, if the state you are restoring
to has a non-directory at the same path as the directory.

Introduce a safety valve to fail the operation in such case, using
the "ls-files --killed" which was designed for this exact purpose.

The "stash save" is stopped when untracked files need to be
discarded because their leading path ceased to be a directory, and
the user is required to pass --force to really have the data
removed.

Signed-off-by: Petr Baudis <pasky@ucw.cz>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 72a08806569c51e1e840901dc8395e0538699ef0 [formerly 72a08806569c51e1e840901dc8395e0538699ef0 [formerly a73653130edd6a8977106d45a8092c09040f9132]]
Former-commit-id: 9a1cf82501947c97c32dba93c7183b3013f5791c
Former-commit-id: 367e5baa8d84915db6b386929bb004ce2661d767