fixup-builtins: retire an old transition helper script
This script was added in 36e5e70 (Start deprecating "git-command" in
favor of "git command", 2007-06-30) with the intent of aiding the
transition away from dashed forms.

It has already been used to help the transision and served its
purpose, and is no longer very useful for follow-up work, because
the majority of remaining matches it finds are false positives.

Signed-off-by: Ramkumar Ramachandra <artagnon@gmail.com>
Reviewed-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 59ae30984e830ea9cf468eae1f5b034420dae8de [formerly 59ae30984e830ea9cf468eae1f5b034420dae8de [formerly 531c8dd4fb18e653d6170d367e4b23c3164d5ce0]]
Former-commit-id: dff31086319ffd0e02f7ec31d0d02b9053f55f0e
Former-commit-id: 783f87f0a3f5293e5cf6eb090af20e288dcb431d