merge: handle --ff/--no-ff/--ff-only as a tri-state option
These three options mean "favor fast-forwarding when possible,
without creating an unnecessary merge", "never fast-forward and
always create a merge commit even when the commit being merged is a
strict descendant", and "we do not want to create any merge commit;
update only when the merged commit is a strict descendant".

They are "pick one out of these three possibilities" options, and
correspond to "merge.ff" configuration that is tri-state (yes, no
and only).

However, the implementation did not follow the usual convention for
the command line options (later one wins, and command line overrides
what is in the configuration).

Fix this by consolidating two variables (fast_forward_only and
allow_fast_forward) used in the implementation into one enum that
can take one of the three possible values.

Signed-off-by: Miklos Vajna <vmiklos@suse.cz>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 6c7c07e2f83e0bbdee22f3fc3c73e1cc8afa04aa [formerly 927aeec0709854dddccf63c976d11d53e3e90440] [formerly 927aeec0709854dddccf63c976d11d53e3e90440 [formerly a54841e96b78203598dea6b31b9618f40f107e7b]]
Former-commit-id: 1854243898a65dc9376eb7c4210ac2a3dd28c134
Former-commit-id: dcac8b4114b895274b440651d1135f4b3d515106