Add --depth to submodule update/add
Add the --depth option to the add and update commands of "git submodule",
which is then passed on to the clone command. This is useful when the
submodule(s) are huge and you're not really interested in anything but
the latest commit.

Tests are added and some indention adjustments were made to conform to the
rest of the testfile on "submodule update can handle symbolic links in pwd".

Signed-off-by: Fredrik Gustafsson <iveqy@iveqy.com>
Acked-by: Jens Lehmann <Jens.Lehmann@web.de>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 0e07b4dd922368fdd157fcbf0d3cbd34484181c0 [formerly 02d43e1f725f25c16af59be49d2f4f4bb5170583] [formerly 02d43e1f725f25c16af59be49d2f4f4bb5170583 [formerly 275cd184d52b5b81cb89e4ec33e540fb2ae61c1f]]
Former-commit-id: f0950fdfef612f308e0ea9b3cb30e6fc4d21ecc7
Former-commit-id: 587db36ffa83dfc53e2ff24c9382a417a63ea8f8