submodule update: allow custom command to update submodule working tree
Users can set submodule.$name.update to '!command' which will cause
'command' to be run instead of checkout/merge/rebase. This allows
the user finer-grained control over how the update is done.

The primary motivation for this was interoperability with stgit;
however being able to intercept the submodule update process may
prove useful for integrating with or extending other tools.

Signed-off-by: Chris Packham <judge.packham@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 852372181a6de6c055fcd1d8ccfd945069b50ee3 [formerly 852372181a6de6c055fcd1d8ccfd945069b50ee3 [formerly 6cb5728c43f34a7348e128b44b80d00b9417cb19]]
Former-commit-id: 61d11e86483da62f2b1539945760fb48272e43ca
Former-commit-id: 07eff127adbc109141ba9ee38324e2bd0089f463