git-remote-mediawiki: introduction of Git::Mediawiki.pm
We would want to allow the user to preview what he has edited locally
before pushing it out (and thus creating a non-removable revision in
the mediawiki's history).

This patch introduces a new perl package in which we will be able to
share code between that new tool and the remote helper:
git-remote-mediawiki.perl.

A perl package offers the best way to handle such case: Each script
can select what should be imported in its namespace.  The package
namespacing limits the use of side effects in the shared code.

An alternate solution is to concatenate a "toolset" file with each
*.perl when 'make'-ing the project. In that scheme, everything is
imported in the script's namespace. Plus, files should be renamed in
order to chain to Git's toplevel makefile. Hence, this solution is not
acceptable.

Signed-off-by: Benoit Person <benoit.person@ensimag.fr>
Signed-off-by: Matthieu Moy <matthieu.moy@grenoble-inp.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: a98ac4d5fa12f8aa46e8a7e833dce5acdea2a800 [formerly a98ac4d5fa12f8aa46e8a7e833dce5acdea2a800 [formerly e19189060fcfb62410ce8e008720dbb6e7fa0d7b]]
Former-commit-id: 0e29c0e3673ddd5ae58cc278e417bd563b39bdc1
Former-commit-id: 5c4a983b28d45977fe76ef594835bc25876b77a1