wrap-for-bin: make bin-wrappers chainable
For now, bin-wrappers overwrites GITPERLLIB. If we want to chain to
those scripts and define GITPERLLIB before, our changes will be
discarded.

This patch makes the bin-wrappers prepend their modifications to
GITPERLLIB rather than redefining it. It also unset GITPERLLIB in the
test-suite to prevent broken $GITPERLLIB in the user's configuration
from interfering with the testsuite.

The codes using GIT_TEMPLATE_DIR and GIT_TEXTDOMAINDIR handle only one
path in each of this variable so this new behavior would be useless on
those variables.

Signed-off-by: Benoit Person <benoit.person@ensimag.fr>
Signed-off-by: Matthieu Moy <matthieu.moy@grenoble-inp.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: f3cf8a1fc5261a5bcdb8b398c3f36b4948a2d69a [formerly f3cf8a1fc5261a5bcdb8b398c3f36b4948a2d69a [formerly 8bade1e12e2041a21f50009d8baaa11bcb800e8f]]
Former-commit-id: 79f589f438a42f24442c617c817333b0ae50abe2
Former-commit-id: ba7e3555e3a66f324fabd579d189efe0fbe07b5d