git-remote-mediawiki: add a git bin-wrapper for developement
The introduction of the Git::Mediawiki package makes it impossible to test,
without installation, git-remote-mediawiki and git-mw.

Using a git bin-wrapper enables us to define proper $GITPERLLIB to force the
use of the developement version of the Git::Mediawiki package, bypassing its
installed version if any.

An alternate solution was to 'install' all the files required at each build
but it pollutes the toplevel with untracked files.

Signed-off-by: Benoit Person <benoit.person@ensimag.fr>
Signed-off-by: Matthieu Moy <matthieu.moy@grenoble-inp.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 3e3f27dc1bd19af4b93716ff9123afeed40360d7 [formerly 1effc2ba649bb4837deb3b03271375cadc7ea592] [formerly 1effc2ba649bb4837deb3b03271375cadc7ea592 [formerly 8a7c21514073f25e8262e311c0fa93101da0e6d3]]
Former-commit-id: b1f1b19f47096c444ee6ddb26c5c4588b17c0eb8
Former-commit-id: 92730604f4c15f7cf27ccfc56d5e5a210b8971f3