git-remote-mediawiki: add git-mw command
For now, git-remote-mediawiki is only a remote-helper. This patch adds a new
toolset script in which we will be able to build new tools for
git-remote-mediawiki.

This toolset uses a subcommand-mechanism to launch the proper action. For now
only the 'help' subcommand is implemented. It also provides some generic code
for the verbose and help command line options.

Signed-off-by: Benoit Person <benoit.person@ensimag.fr>
Signed-off-by: Matthieu Moy <matthieu.moy@grenoble-inp.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: d665519615bb3a67f30e5c3437b45e9ce62c3221 [formerly a81d1421b597a3bc0bd568b28ee1fb5fd75289dc] [formerly a81d1421b597a3bc0bd568b28ee1fb5fd75289dc [formerly 07a263b9056ac7135ed84d64cabf80c5a6b1972e]]
Former-commit-id: e605d180a388e1e485fcea8036fd5a6edf0c36d1
Former-commit-id: 8f6af14052a0bb266b73fa8d6796c9c072235898