git-remote-mediawiki: add preview subcommand into git mw
In the current state, a user of git-remote-mediawiki can edit the markup text
locally, but has to push to the remote wiki to see how the page is rendererd.
Add a new 'git mw preview' command that allows rendering the markup text on
the remote wiki without actually pushing any change on the wiki.

This uses Mediawiki's API to render the markup and inserts it in an actual
HTML page from the wiki so that CSS can be rendered properly. Most links
should work when the page exists on the remote.

Signed-off-by: Benoit Person <benoit.person@ensimag.fr>
Signed-off-by: Matthieu Moy <matthieu.moy@grenoble-inp.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: f612a05a86ab92a6e29f81686665721e4544c4b1 [formerly f612a05a86ab92a6e29f81686665721e4544c4b1 [formerly 0078a7fa05b3d671739aefba5265da809ea32e55]]
Former-commit-id: 7e5a2dbefa954eb11bd2ec4245467b94a33c9937
Former-commit-id: 6082e21f8347bad87564d8487a0861abae4aa920