zero-initialize object_info structs
The sha1_object_info_extended function expects the caller to
provide a "struct object_info" which contains pointers to
"query" items that will be filled in. The purpose of
providing pointers rather than storing the response directly
in the struct is so that callers can choose not to incur the
expense in finding particular fields that they do not care
about.

Right now the only query item is "sizep", and all callers
set it explicitly to choose whether or not to query it; they
can then leave the rest of the struct uninitialized.

However, as we add new query items, each caller will have to
be updated to explicitly turn off the new ones (by setting
them to NULL).  Instead, let's teach each caller to
zero-initialize the struct, so that they do not have to
learn about each new query item added.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 0cae94d100c7b60ad15505273b6403838d572951 [formerly 0cae94d100c7b60ad15505273b6403838d572951 [formerly 7c07385d902f6d8c177d533dc2faa36ef4a52a66]]
Former-commit-id: e09e97f121e2f21192931606078b53c9048796c4
Former-commit-id: 39ee9d5631f5bb6ef6f0f6615737b87eaa93cc7d