teach sha1_object_info_extended a "disk_size" query
Using sha1_object_info_extended, a caller can find out the
type of an object, its size, and information about where it
is stored. In addition to the object's "true" size, it can
also be useful to know the size that the object takes on
disk (e.g., to generate statistics about which refs consume
space).

This patch adds a "disk_sizep" field to "struct object_info",
and fills it in during sha1_object_info_extended if it is
non-NULL.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 18df84c4146db9ffb016202f40e0ae16e04c17dd [formerly ac2dbd214878f7a0d4a575a2882fe2ee26e0a97a] [formerly ac2dbd214878f7a0d4a575a2882fe2ee26e0a97a [formerly 161f00e708874bac646da2ac05c66a18ade2c074]]
Former-commit-id: 3ff9f9d6a9a29c3c35c40623f89c0b476a7eb6b4
Former-commit-id: 50b5dd9a4ec50f773f9bff27d3b3f119861cc8c7