push: avoid suggesting "merging" remote changes
With some workflows, it is more suitable to rebase on top of remote
changes when a push does not fast-forward.  Change the advice messages
in git-push to suggest that a user "integrate the remote changes"
instead of "merge the remote changes" to make this slightly clearer.

Also change the suggested 'git pull' to 'git pull ...' to hint to users
that they may want to add other parameters.

Suggested-by: Philip Oakley <philipoakley@iee.org>
Signed-off-by: John Keeping <john@keeping.me.uk>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 8d7e2df5a37dc9a14f169bc0ea9ad8aa5208b7ec [formerly 8d7e2df5a37dc9a14f169bc0ea9ad8aa5208b7ec [formerly fc6c4e96f13fe595fc8991d91c56f2abbe67dc89]]
Former-commit-id: 7c6baea0554b945b4e8d9ad2740ea425ed3676b8
Former-commit-id: a8dfbea9e3e2e00d7753e6e80c8e35fa1a6579eb