name-rev: allow converting the exact object name at the tip of a ref
"git name-rev" is supposed to convert given object names into
strings that name the same objects based on refs, that can be fed to
"git rev-parse" to get the same object names back, so the output for
the commit object v1.8.3^0 (i.e. the commit tagged as v1.8.3)

    $ git rev-parse v1.8.3 v1.8.3^0 | git name-rev --stdin
    8af06057d0c31a24e8737ae846ac2e116e8bafb9
    9e7f22c269e3ff87a8771c62d5a912aca71c0cb6 [formerly 4790bf8fa23c45f8ea9deddf88ae2941479a2665] [formerly 8b093e3dd49878e3ed2de59b1594f49d16a84605] [formerly 1d064cae2c6d36a2bf6364829af86ead6d34dd26 [formerly 1d064cae2c6d36a2bf6364829af86ead6d34dd26 [formerly edca4152560522a431a51fc0a06147fc680b5b18]]] (tags/v1.8.3^0)

has to have "^0" at the end, as "edca41" is a commit, not the tag
that references it.  But we do not get anything for the tag object
(8af0605) itself.

This is because the command however did not bother to see if the
object is at the tip of some ref, and failed to convert a tag
object.

Teach it to show this instead:

    $ git rev-parse v1.8.3 v1.8.3^0 | git name-rev --stdin
    8af06057d0c31a24e8737ae846ac2e116e8bafb9 (tags/v1.8.3)
    9e7f22c269e3ff87a8771c62d5a912aca71c0cb6 [formerly 4790bf8fa23c45f8ea9deddf88ae2941479a2665] [formerly 8b093e3dd49878e3ed2de59b1594f49d16a84605] [formerly 1d064cae2c6d36a2bf6364829af86ead6d34dd26 [formerly 1d064cae2c6d36a2bf6364829af86ead6d34dd26 [formerly edca4152560522a431a51fc0a06147fc680b5b18]]] (tags/v1.8.3^0)

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 6c6ed7c81ac2d5e693b00dfa22f620b36ba252e5 [formerly 6c6ed7c81ac2d5e693b00dfa22f620b36ba252e5 [formerly b23e0b9353e38149e791bde0feb1220aa1c8bac1]]
Former-commit-id: 75fa639a3a3b4d0d56d0d521988132c2680b4447
Former-commit-id: 0e8eb5ac7bf9b54f223ee72e0583ddc6576556f8