cache.h: move remote/connect API out of it
The definition of "struct ref" in "cache.h", a header file so
central to the system, always confused me.  This structure is not
about the local ref used by sha1-name API to name local objects.

It is what refspecs are expanded into, after finding out what refs
the other side has, to define what refs are updated after object
transfer succeeds to what values.  It belongs to "remote.h" together
with "struct refspec".

While we are at it, also move the types and functions related to the
Git transport connection to a new header file connect.h

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: c6f103144c5aba49180c86f51388770c1b0a6a76 [formerly 3dd4a4ddab7b392ac02a2e6211eac864c1b6206d] [formerly 3dd4a4ddab7b392ac02a2e6211eac864c1b6206d [formerly 47a59185369b8905ad3a4012688cba92fd2ac1ff]]
Former-commit-id: cc583a310ceb855ca1105ceb86d60ee7b28c1c10
Former-commit-id: 1f73af86089edcc220fe80832ce71f250ff77ea6